refactor, use older matplotlib keyword
