+++ infra/platforms/platforms/numpy.py
@@ -43,7 +43,7 @@ class NumpyRunner(Runner):
 
         for core, driver_dir in zip(cores, driver_dirs):
             driver_path = os.path.join(driver_dir, driver_name)
-            _, sample = self.cache.get_core(core.key)
+            sample = self.cache.get_sample(core.key, self.seed)
             input_points, _ = sample
             with open(driver_path, 'w') as f:
                 print('import time', file=f)
@@ -83,6 +83,8 @@ class NumpyRunner(Runner):
         # run processes sequentially
         times = [[] for _ in driver_dirs]
         for i, driver_dir in enumerate(driver_dirs):
+            log_prefix = f'[{i}/{len(driver_dirs)}] '
+            print(log_prefix, end='', flush=True)
             for _ in range(self.num_runs):
                 driver_path = Path(os.path.join(driver_dir, driver_name))
                 p = Popen([target, driver_path], stdout=PIPE)
@@ -93,8 +95,11 @@ class NumpyRunner(Runner):
                     raise RuntimeError('Unexpected error when running {out_path}: {output}')
                 times[i].append(float(time.group(1)))
                 print('.', end='', flush=True)
-            print('x', end='', flush=True)
-        print()
+
+            # Reset terminal
+            print('\r', end='', flush=True)
+            print(' ' * (len(log_prefix) + self.num_runs), end='', flush=True)
+            print('\r', end='', flush=True)
         
         times = [sum(ts) / len(ts) for ts in times]
         self.log(f'run drivers')
